+++ align_example.py
@@ -1,6 +1,6 @@
-from aligner import SentenceAligner
+import simalign
 
-model = SentenceAligner()
+model = simalign.SentenceAligner()
 result = model.get_word_aligns(["Sir Nils Olav III. was knighted by the norwegian king .",
                        "Nils Olav der Dritte wurde vom norwegischen König zum Ritter geschlagen ."])
 print(result)
