+++ simalign/setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup
+
+setup(name='simalign',
+      version='0.1',
+      description='Word Alignments using Pretrained Language Models',
+      url='https://github.com/masoudjs/simalign',
+      author='Masoud Jalili Sabet, Philipp Dufter',
+      author_email='philipp@cis.lmu.de,masoud@cis.lmu.de',
+      license='GPL-3.0',
+      packages=['simalign'],
+      zip_safe=False)
